requirements.txt
pyyaml
requests
pyperclip
keyboard
python-dotenv
